requirements-postgres.txt
psycopg2-binary==2.9.6
python-dotenv==1.0.0
tqdm==4.65.0
sqlalchemy>=2.0.0
asyncpg
greenlet